refactor: assignment type 경로 변경 및 중복 파일 삭제
