+++ scripts/common/type/assignment.type.js
